% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP_inflator}
\alias{GDP_inflator}
\title{DATASET_TITLE}
\format{
A data frame with 17 rows and 2 variables:
\describe{
  \item{\code{Annual}}{double COLUMN_DESCRIPTION}
  \item{\code{Annual value}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GDP_inflator
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
