% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdpdef_stfed}
\alias{gdpdef_stfed}
\title{DATASET_TITLE}
\format{
A data frame with 90 rows and 2 variables:
\describe{
  \item{\code{YEAR}}{double COLUMN_DESCRIPTION}
  \item{\code{GDPDEF}}{double COLUMN_DESCRIPTION} 
}

A data frame with 90 rows and 2 variables:
\describe{
  \item{\code{YEAR}}{double COLUMN_DESCRIPTION}
  \item{\code{GDPDEF}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
gdpdef_stfed

gdpdef_stfed
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
