% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GearType_NE}
\alias{GearType_NE}
\title{DATASET_TITLE}
\format{
A data frame with 15 rows and 19 variables:
\describe{
  \item{\code{Gear}}{character COLUMN_DESCRIPTION}
  \item{\code{Lobster}}{double COLUMN_DESCRIPTION}
  \item{\code{Herring}}{double COLUMN_DESCRIPTION}
  \item{\code{Sea Scallop}}{double COLUMN_DESCRIPTION}
  \item{\code{Summer Flounder}}{double COLUMN_DESCRIPTION}
  \item{\code{Black Sea Bass}}{double COLUMN_DESCRIPTION}
  \item{\code{Scup}}{double COLUMN_DESCRIPTION}
  \item{\code{Bluefish}}{double COLUMN_DESCRIPTION}
  \item{\code{Groundfish}}{double COLUMN_DESCRIPTION}
  \item{\code{Mackerel}}{double COLUMN_DESCRIPTION}
  \item{\code{Monkfish}}{double COLUMN_DESCRIPTION}
  \item{\code{Red Crab}}{double COLUMN_DESCRIPTION}
  \item{\code{Spiny Dogfish}}{double COLUMN_DESCRIPTION}
  \item{\code{Skate}}{double COLUMN_DESCRIPTION}
  \item{\code{Surfclam/Ocean Quahog}}{double COLUMN_DESCRIPTION}
  \item{\code{Whiting/Other Hakes}}{double COLUMN_DESCRIPTION}
  \item{\code{Tilefish}}{double COLUMN_DESCRIPTION}
  \item{\code{Jonah Crab}}{double COLUMN_DESCRIPTION}
  \item{\code{All Other}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GearType_NE
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
