% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Impacts_Sector_Markups}
\alias{Impacts_Sector_Markups}
\title{DATASET_TITLE}
\format{
A data frame with 100 rows and 4 variables:
\describe{
  \item{\code{Jurisdiction}}{character COLUMN_DESCRIPTION}
  \item{\code{Jurisdiction0}}{character COLUMN_DESCRIPTION}
  \item{\code{Stakeholders}}{character COLUMN_DESCRIPTION}
  \item{\code{Markups}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Impacts_Sector_Markups
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
