% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Imports_2015}
\alias{Imports_2015}
\title{DATASET_TITLE}
\format{
A data frame with 24 rows and 8 variables:
\describe{
  \item{\code{State}}{character COLUMN_DESCRIPTION}
  \item{\code{Imports}}{double COLUMN_DESCRIPTION}
  \item{\code{Whole.Prop}}{double COLUMN_DESCRIPTION}
  \item{\code{Proc.Prop}}{double COLUMN_DESCRIPTION}
  \item{\code{Groc.Prop}}{double COLUMN_DESCRIPTION}
  \item{\code{Rest.Prop}}{double COLUMN_DESCRIPTION}
  \item{\code{Exports from US}}{double COLUMN_DESCRIPTION}
  \item{\code{X8}}{logical COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Imports_2015
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
