% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPD2017}
\alias{IPD2017}
\title{DATASET_TITLE}
\format{
A data frame with 89 rows and 4 variables:
\describe{
  \item{\code{Year}}{double COLUMN_DESCRIPTION}
  \item{\code{IPD}}{double COLUMN_DESCRIPTION}
  \item{\code{2008 Base}}{double COLUMN_DESCRIPTION}
  \item{\code{2017 Base}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
IPD2017
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
