% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Multipliers_Commercial}
\alias{Multipliers_Commercial}
\title{DATASET_TITLE}
\format{
A data frame with 2513 rows and 36 variables:
\describe{
  \item{\code{Notes}}{logical COLUMN_DESCRIPTION}
  \item{\code{Jurisdiction}}{character COLUMN_DESCRIPTION}
  \item{\code{Jurisdiction_Long}}{character COLUMN_DESCRIPTION}
  \item{\code{GearType}}{logical COLUMN_DESCRIPTION}
  \item{\code{Stakeholders}}{character COLUMN_DESCRIPTION}
  \item{\code{Index}}{character COLUMN_DESCRIPTION}
  \item{\code{Species}}{character COLUMN_DESCRIPTION}
  \item{\code{RPC-}}{double COLUMN_DESCRIPTION}
  \item{\code{DirectImpacts-PersonalIncomeMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{IndirectImpacts-PersonalIncomeMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{InducedImpacts-PersonalIncomeMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalImpacts-PersonalIncomeMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{DirectImpacts-OutputMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{IndirectImpacts-OutputMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{InducedImpacts-OutputMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalImpacts-OutputMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{DirectImpacts-EmploymentMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{IndirectImpacts-EmploymentMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{InducedImpacts-EmploymentMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalImpacts-EmploymentMultipliers_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{DirectImpacts-PersonalIncomeImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{IndirectImpacts-PersonalIncomeImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{InducedImpacts-PersonalIncomeImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalImpacts-PersonalIncomeImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{DirectImpacts-OutputImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{IndirectImpacts-OutputImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{InducedImpacts-OutputImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalImpacts-OutputImpacts_000_2006}}{double COLUMN_DESCRIPTION}
  \item{\code{DirectImpacts-EmploymentImpacts_Full-part-timejobs}}{double COLUMN_DESCRIPTION}
  \item{\code{IndirectImpacts-EmploymentImpacts_Full-part-timejobs}}{double COLUMN_DESCRIPTION}
  \item{\code{InducedImpacts-EmploymentImpacts_Full-part-timejobs}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalImpacts-EmploymentImpacts_Full-part-timejobs}}{double COLUMN_DESCRIPTION}
  \item{\code{DirectImpacts-ValueAdded}}{double COLUMN_DESCRIPTION}
  \item{\code{IndirectImpacts-ValueAdded}}{double COLUMN_DESCRIPTION}
  \item{\code{InducedImpacts-ValueAdded}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalImpacts-ValueAdded}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Multipliers_Commercial
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
