% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Multipliers_Recreational}
\alias{Multipliers_Recreational}
\title{DATASET_TITLE}
\format{
A data frame with 390 rows and 6 variables:
\describe{
  \item{\code{Area}}{character COLUMN_DESCRIPTION}
  \item{\code{Mode}}{double COLUMN_DESCRIPTION}
  \item{\code{Name0}}{character COLUMN_DESCRIPTION}
  \item{\code{Name}}{character COLUMN_DESCRIPTION}
  \item{\code{Acount}}{double COLUMN_DESCRIPTION}
  \item{\code{X6}}{logical COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Multipliers_Recreational
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
