% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olo_species_codes}
\alias{olo_species_codes}
\title{DATASET_TITLE}
\format{
A data frame with 1071 rows and 4 variables:
\describe{
  \item{\code{AFS_NAME}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{double COLUMN_DESCRIPTION}
  \item{\code{NAME}}{character COLUMN_DESCRIPTION}
  \item{\code{FUS_NAME}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
olo_species_codes
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
