% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProductFlow_BasicData_StateTotals_test}
\alias{ProductFlow_BasicData_StateTotals_test}
\title{DATASET_TITLE}
\format{
A data frame with 25 rows and 2 variables:
\describe{
  \item{\code{State}}{character COLUMN_DESCRIPTION}
  \item{\code{Total Value}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
ProductFlow_BasicData_StateTotals_test
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
