% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProductFlow_BasicData_Sythesis_test}
\alias{ProductFlow_BasicData_Sythesis_test}
\title{DATASET_TITLE}
\format{
A data frame with 6 rows and 7 variables:
\describe{
  \item{\code{ImportsAndDefaultDistribution}}{character COLUMN_DESCRIPTION}
  \item{\code{Wholesalers}}{double COLUMN_DESCRIPTION}
  \item{\code{Processors}}{double COLUMN_DESCRIPTION}
  \item{\code{Retail}}{double COLUMN_DESCRIPTION}
  \item{\code{Restaurants}}{double COLUMN_DESCRIPTION}
  \item{\code{ExportsFromUS}}{double COLUMN_DESCRIPTION}
  \item{\code{X7}}{logical COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
ProductFlow_BasicData_Sythesis_test
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
