% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProductFlow_BasicData_test}
\alias{ProductFlow_BasicData_test}
\title{DATASET_TITLE}
\format{
A data frame with 14 rows and 9 variables:
\describe{
  \item{\code{Source of fish, fish products}}{character COLUMN_DESCRIPTION}
  \item{\code{Source}}{character COLUMN_DESCRIPTION}
  \item{\code{Primary whole- salers/processors}}{double COLUMN_DESCRIPTION}
  \item{\code{Secondary whole- salers/distributors}}{double COLUMN_DESCRIPTION}
  \item{\code{Foodservice/ restaurants}}{double COLUMN_DESCRIPTION}
  \item{\code{Groceries/ retail markets}}{double COLUMN_DESCRIPTION}
  \item{\code{Exports}}{double COLUMN_DESCRIPTION}
  \item{\code{Final consumers}}{double COLUMN_DESCRIPTION}
  \item{\code{Total}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
ProductFlow_BasicData_test
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
