% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProductFlow_synthesis_comm}
\alias{ProductFlow_synthesis_comm}
\title{DATASET_TITLE}
\format{
A data frame with 18 rows and 8 variables:
\describe{
  \item{\code{Product flow for shrimp and non-shrimp fish and fish products: synthesis of basic data Source of fish, fish products}}{character COLUMN_DESCRIPTION}
  \item{\code{Destination of fish, fish products (percentage distribution) (Note 5) Primary processors}}{double COLUMN_DESCRIPTION}
  \item{\code{Secondary whole- salers/distributors}}{double COLUMN_DESCRIPTION}
  \item{\code{Foodservice/ restaurants}}{double COLUMN_DESCRIPTION}
  \item{\code{Groceries/ retail markets}}{double COLUMN_DESCRIPTION}
  \item{\code{Exports (Notes 6, 7)}}{double COLUMN_DESCRIPTION}
  \item{\code{Final consumers}}{double COLUMN_DESCRIPTION}
  \item{\code{Total}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
ProductFlow_synthesis_comm
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
