% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recstatecodes1}
\alias{recstatecodes1}
\title{DATASET_TITLE}
\format{
A data frame with 25 rows and 5 variables:
\describe{
  \item{\code{state.no}}{double COLUMN_DESCRIPTION}
  \item{\code{state.name}}{character COLUMN_DESCRIPTION}
  \item{\code{region.name}}{character COLUMN_DESCRIPTION}
  \item{\code{region.no}}{double COLUMN_DESCRIPTION}
  \item{\code{state.abb}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
recstatecodes1
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
