% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/statecodes.R
\docType{data}
\name{statecodes}
\alias{statecodes}
\title{DATASET_TITLE}
\format{
A data frame with 25 rows and 4 variables:
\describe{
  \item{\code{State.no}}{double COLUMN_DESCRIPTION}
  \item{\code{State}}{character COLUMN_DESCRIPTION}
  \item{\code{Region}}{character COLUMN_DESCRIPTION}
  \item{\code{Region.no}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
statecodes

statecodes(st, methods.Comm.States, output = "Number")
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
DATASET_DESCRIPTION

This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
\keyword{datasets}
