% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statereg}
\alias{statereg}
\title{DATASET_TITLE}
\format{
A data frame with 25 rows and 11 variables:
\describe{
  \item{\code{X1}}{double COLUMN_DESCRIPTION}
  \item{\code{State}}{character COLUMN_DESCRIPTION}
  \item{\code{State1}}{character COLUMN_DESCRIPTION}
  \item{\code{fips}}{double COLUMN_DESCRIPTION}
  \item{\code{Region}}{character COLUMN_DESCRIPTION}
  \item{\code{abbvst}}{character COLUMN_DESCRIPTION}
  \item{\code{abbvreg}}{character COLUMN_DESCRIPTION}
  \item{\code{xstate}}{double COLUMN_DESCRIPTION}
  \item{\code{xreg}}{double COLUMN_DESCRIPTION}
  \item{\code{State.no}}{double COLUMN_DESCRIPTION}
  \item{\code{Region.no}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
statereg
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
